requirements.txt
openai
streamlit
python-dotenv
streamlit-mic-recorder
streamlit-audiorecorder
faiss-cpu 
numpy 